Hook Improvements. See Details
Hook now wraps around objects (mostly) correctly.

Nodes in hook are now stored in a C# list instead of a linked list.

Whenever a new node is created, it will check if it is in los of the node before the last to see if the node between them is necessary.