Rope now wraps. See Details
Pretty proud of this one. Rope now properly wraps around objects (though it doesn't presently unwrap). Created a cylinder within the platforms. The point where the rope collides with an object casts a line towards the cylinder and gets the normal vector from it. That normal vector is used as the offset for node so it continues to collide properly.